Expand single-child folders in the new tree view (#47117)
Part of #46602

This adds single-child folder expansion to the new tree view based on
how VS Code does it 🎉

Since the implementation of the tree is append-only, only a bit of
bookkeeping is needed to detect single-child expansion scenarios and
change the tree entries accordingly. I've left various comments in the
code which should explain the implementation for the curious.