redispool: use postgres for redispool.Store in App
We introduce a new table "redis_key_value" which is written to by
KeyValue. The DB backed KeyValue store needs to wait for the database to
be ready. However, Store is a package global. As such we use a pattern
were we register the DB into redispool once the DB is ready. Before that
all the KeyValue operations will return an error.

Test Plan: go test and manual testing of Sourcegraph App.