experimental seach input: Add back toggle buttons
This PR adds back the toggle buttons to the search input. This is an
intermediate solution until we've explored whether we can provide the
same functionality in a different way.

Most notable difference is that the toggles are only shown when the
input is focused. Unfortunately this doesn't work well with the smart
search toggle button popover. Opening the popover will hide the buttons.
A couple of ways to address this:
- Don't show the smart search toggle button on the homepage.
- Always show toggle buttons (not only on focus)
- Find a way to render the popover inside the the search input container