singleprogram: use localhost for gitserver
Using the hostname isn't gaurenteed to work, it depends on how the
machine is configured. For example on my linux desktop this fails. We
only start the services on localhost and we use ipv4 localhost address
for all other services.

Test Plan: sg start app doesn't spam my logs with failing to connect to
gitserver.