singleprogram: use localhost for gitserver (#47366)
* singleprogram: use localhost for gitserver

Using the hostname isn't gaurenteed to work, it depends on how the
machine is configured. For example on my linux desktop this fails. We
only start the services on localhost and we use ipv4 localhost address
for all other services.

Additionally introduce GITSERVER_EXTERNAL_ADDR. Treating the hostname as
an address is incorrect. We instead use a more common pattern of having
an environment variable which specifies the reachable address for the
gitserver service.

Test Plan: sg start app and sg start work