Ranking: remove hoistHighScore, since it's no longer needed (#47356)
Now that we incorporate file ranking signals into the main score, this logic
doesn't have any effect. This is because the top-scoring result is already at
the top.

If we find we still need to boost results with a high query match score, we'll
do that a different way (by decreasing the weight of file ranks).