Search: fix zoekt result limits (#47343)
For indexed search, frontend translates a user's query into zoekt options. This
PR makes these fixes to the zoekt result limits, which could help search
performance and memory usage a bit:
* Removes `resultCountFactor`, and instead uses simple default limits for
ShardMaxDocCount and TotalMaxDocCount.
* Updates old logic that increases ShardMaxDocCount and TotalMaxDocCount too
much. The check used the wrong default value, so it always fired and multiplied
the intended value by 50.
* Removes MaxDocDisplayCount adjustment. Before we always added 2000, but this
is not needed because we no longer use these extra results.