searcher: use the same gitserver fetch timeout as symbols
Searcher has a hardcoded timeout of 10 minutes. Interestingly symbols
has a hardcoded timeout of 2 hours. Now this is likely due to fetching
more history for a repository for rockskip. However, I suspect 10
minutes is too aggressive for large monorepos. As such we use a much
more conservative timeout of 2h.

The downside of this approach is we have a relatively naive semaphore
controlling concurrent fetches. So if we have a large burst of activity
that never gets retried it will take a lot longer for us to cancel out
the backlog. However, in practice this feels like the better tradeoff as
well as this being consistent with symbols.

Test Plan: go test