avoid duplicate but conflicting PROCESSING_TIMEOUT env var registrations (#47513)
When running App, where all env vars are applied to the same process,
the 2 env var `PROCESSING_TIMEOUT` definitions caused a panic:

```
panic: env var "PROCESSING_TIMEOUT" already registered with a different description or value
```

The values were the same, but the canonical string representation of
`2h` as a Go `time.Duration` is actually `2h0m0s`. This change causes it
not to panic.




## Test plan

Run `sg start app` and see it not panic.