RFC rcache: use context for FIFOList.Insert
We want to be able to timeout inserts just in case redis is acting slow.
However, for every call-site to Insert we probably want the same
behaviour:

- Ignore passed in ctx (it may be canceled).
- Run in background.
- Use a timeout.

This makes me think this design is the incorrect one, but I am sending
it out anyways to spark discussion.

I think a nicer approach is instead for FIFOList to have a configurable
timeout and not take in a context. Since that will be a footgun. It may
be nicer to wrap FIFOList with something like "BackgroundRecorder" or
something which does what we expect, and leave FIFOList to be a nice
pure FIFOList.

WDYT?

Test Plan: CI