[fix] Better UX for test connection button
Previously, the behavior was not showing enough feedback
that there is an ongoing process running in the background. Added
a loading spinner on the button and a message below it saying,
that we load data.

Also, previously we did not handle errors correctly. Now we show
an error alert even if there is an error communicating to the graphql
backend or some other unexpected problem.

Related: https://github.com/sourcegraph/sourcegraph/issues/47522