[fix] Better UX for test connection button (#47539)
## Description

Previously, the behavior was not showing enough feedback that there is
an ongoing process running in the background. Added a loading spinner on
the button and a message below it saying, that we load data.

Also, previously we did not handle errors correctly. Now we show an
error alert even if there is an error communicating to the graphql
backend or some other unexpected problem.

Related: https://github.com/sourcegraph/sourcegraph/issues/47522

## Video

https://www.loom.com/share/e6a09b86a2bb4c1d9e39c3274cacdce5

## Test plan

Tested locally with various failure modes:
- connection succeeds
- connection does not succeed because of bad URL for code host
- connection does not succeed because of bad token for code host
- kill frontend - graphql request fails

I was not able to reproduce the redis problem locally though, even when
I used the suggested settings.

## App preview:

- [Web](https://sg-web-milan-test-connection.onrender.com/search)

Check out the [client app preview
documentation](https://docs.sourcegraph.com/dev/how-to/client_pr_previews)
to learn more.