insights: batch backfilling in groups of repos  (#47381)
Because some repos take longer than other to backfill, the throughput is
not always maximized, by grabbing a batch of repos at a time we can run
can continue to run additional repos though the backfill process while the
larger/slower repo is running.  

This sets a conservative default page size of 10 and limits the concurrency to 3 repos backfilling at once.  Previous rate limits for searches per second still apply. A batch or repos will complete as a group, this means that the runtime of an iteration could exceeded the configured time, by an extended period of a batch contained multiple large repos.  This should not be problematic because if this insight is backfilling it is was the highest priority.


## Test plan
Existing unit test coverage
Created and backfilled a new insight
<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->