add a basic Cody page at /cody (feature-flagged)
This adds a basic page that lets users interact with Cody. It is just a single-chat-message interface and does not include any of the special Cody context or other features yet. It's mainly intended to make it easier for others to hack on these in the web app.

See https://docs.google.com/document/d/1u7HYPmJFtDANtBgczzmAR0BmhM86drwDXCqx-F2jTEE/edit# for instructions.

Actual Cody functionality in the web app will likely be done in a separate (private) repository and pulled in at build time into this repository.