experimental search input: Hide copy and smart search buttons on homepage (#47529)
To make the input less "noisy" on the homepage we are hiding the smart
search and copy query buttons. The assumption is that those buttons are
less relevant in this context (they are more relevant *after* you
performed a query). The buttons will be available on the search results
page (the experimental search input will be enabled on the search
results page in another PR).