searcher: expose CacheState API via grpc only
The intention of the API is for Code Insights to intelligently ask
searcher what the state of fetching from gitserver is. We only expose
this via grpc to simplify implementation + we only intend on targetting
a release of this API in the March release (enough time for grpc to be
ready?)

This is just WIP, because I realised some complications when
implementing. The first of which being that hybrid search won't go via
the usual cache code path, it will first speak to zoekt and gitserver.

This makes me think the design of using diskcache is incorrect, and
rather the searcher service itself needs to do bookkeeping around if it
has managed to do a search for a repo@commit yet/etc. But I am sending
this out anyways, since the grpc API won't be that different so it can
be reviewed.

Note: I don't think you can directly use the grpc client, since we need
to do the same routing we do for search requests. That is not
implemented yet either.

Test Plan: unit tests added. Will also test manually via the grpc debug
endpoints.