🌪️ Add gating and fork web app entrypoints (#47611)
- We prefix existing components for both enterprise and open-source with
`Legacy*` (e.g. `LegacySourcegraphWebApp`, `LegacyLayout`). This is to
denote that we do not plan to update these components anymore.

- It creates a site-config level feature flag for Storm. Why site-level?
Mostly because this way, we avoid any lazy evaluation of feature flags
and can branch very early based on the `window.context`. This is
necessary since neither the feature testing infra nor the settings infra
is initialized when we need to make the switch. To enable them, do
something like this in `dev-private`:
  
  ```diff
  +++ b/enterprise/dev/site-config.json
  @@ -52,7 +52,8 @@
      "openTelemetry": { "endpoint": "/-/debug/otlp" }
    },
    "experimentalFeatures": {
  -    "enableLegacyExtensions": false
  +    "enableLegacyExtensions": false,
  +    "enableStorm": true
    },
    "externalService.userMode": "all",
    "outboundRequestLogLimit": 50
  ```

- I’m then moving on to forking `Layout` and `SourcegraphWebApp`. The
latter needed to be converted from a class component to a function
component. We're still loading the same legacy components. The new `Layout` 
no longer renders the React router but instead uses the `<Outline />` pattern. 
The React router entrypoint is moved into `SourcegraphWebApp` and
now uses route children for the first time.