monitoring: make traffic graph top row for searcher
We introduce another graph which is the rate of traffic per response
code. This is the first graph seen and gives a much better picture of
what the service is doing and how well it is doing it.

We keep the error percantage graph, but now it is the second row. The
next commit will add another graph on that row to make it less big.