reference panel: Don't error when language spec cannot be found
I don't know what the original intend behind throwing an error was, but
from a type safety perspective it seems better to return
`undefined`/`null` when the spec is not available and let the UI handle
the case.

In this PR I'm simply rendering the same UI as if the token wasn't found.