insights: all repo iterator to pull multiple repos at a time (#47352)
Adds new capability to the repo iterator to allow a consumer to pull
multiple repos at a time and then call Finish once to complete them all.

Uses this new iterator capability because some repos take longer than
other to backfill, the throughput is not always maximized. By grabbing a
batch of repos at a time we can run can continue to run additional repos
though the backfill process while the larger/slower repo is running.

This sets a conservative default page size of 10 and limits the
concurrency to 3 repos backfilling at once. Previous rate limits for
searches per second still apply. A batch of repos will complete as a
group, this means that the runtime of an iteration could exceeded the
configured time, by an extended period of time if a batch contained multiple
large repos. This should not be problematic because if this insight is
backfilling, it is was the highest priority.