gRPC: prep for converting gitserver.Search (#47584)
This is just some prep work for converting the `gitserver.Search`
endpoint. It comes in the form of three commits:
1) Make `Server.search()` HTTP-agnostic. Previously, it took a JSON
buffer as an argument, but now it takes a match callback that can be
used for both gRPC and HTTP.
2) Move all the observability code into a shareable wrapper. This keeps
all the observability code still somewhat separate from the search code,
but makes it possible to use it from both gRPC and HTTP calls.
3) Rename the `internal/gitserver/proto` package to
`internal/gitserver/v1`. I had thought we already did this, but
apparently not.