wrap the correct error in the gitserver client (#47697)
We were wrapping `err`, which is guaranteed to be `nil`.
Wrapping a `nil` error returns `nil`. Instead, I expect the intent here
was to wrap `filteringErr`, which is guaranteed to be non-nil here. I
expect that this was causing an issue where we would return a `nil` 
`fs.FileInfo`, but also a `nil` error,  which would cause a panic further 
up the line when a caller tried to use `fs.FileInfo`.