fix goroutine leak for comby streaming
This fixes an issue with the code that streams to comby where an error
from comby would cause the writer to block indefinitely because the
channel it was writing to was no longer being read from. Additionally,
it propagates the error from comby rather than just logging it, and
converts the goroutine spawning to use conc/pool for clarity on errors
and panic safety.