web: fix aggregation search bar click (#47719)
An alternative fix for the issue described in this PR
- https://github.com/sourcegraph/sourcegraph/pull/47713

We discussed the problem with Vova on the Zoom call. The root cause of
the issue is that previously we used mutable location object that didn't
depend on component re-renders which is not the case anymore. That's why
the `submitSearch` handler uses the outdated search query.

## Test plan

- Go to the home page (search any query with capture group param, for
example, `file:go\.mod go\s*(\d\.\d+) [\s\S]*
github.com/sourcegraph/log`
- You should see that by default, we provide capture group aggregation 
- If you pick one of the aggregation bars, you probably should see the
other aggregation type (not capture group since we replace the original
query with a specific bar capture group value)
- In case of `file:go\.mod go\s*(\d\.\d+) [\s\S]*
github.com/sourcegraph/log` it should be repo aggregation