simplify blame streamer
Streaming git blame was not closing the `io.ReadCloser` it takes
ownership of. This adds a `Close()` method to the `HunkReader` type that
cleans up its resources.

Because the `ReadCloser` is passed into a spawned goroutine, I can't
close it without both holding a reference both from the reader and the
spawned goroutine, and waiting for the goroutine to exit first, which
was not easy to do. Instead, I did a small refactor that avoids spawning
a new goroutine, preferring to lazily parse the stream. This has the
added benefit of being simpler to reason about.