Search: fix goroutine leak for comby streaming (#47718)
This fixes an issue with the code that streams to comby where an error
from comby would cause the writer to block indefinitely because the
channel it was writing to was no longer being read from. Additionally,
it propagates the error from comby rather than just logging it, and
converts the goroutine spawning to use conc/pool for clarity on errors
and panic safety.

I noticed the goroutine leak while investigating [a separate goroutine
leak](https://github.com/sourcegraph/sourcegraph/pull/47715).

## Test plan

Added a fragile unit test to ensure that an error from comby does not
cause the call to block indefinitely. Integration tests are running.