bazel: update buildfiles (#47744)
Follow-up to https://github.com/sourcegraph/sourcegraph/pull/47540 which
updated the grpc gateway and required to update the patch.

It also adds the necessary `#keep` comments to avoid having gazelle
messing up the current buildfiles.

Release: this doesn't affect the app in anyway, so we're safe. 

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

```
bazel build //internal/...
``` 

suceeded.