experimental search input: Fix context filter handling
This commit fixes an issue with the context filter on the search result
and repo pages.
When we sync the query from the URL to the input we've been removing the
context filter from the query, because we currently have a separate
context selector. The new input doesn't have that though so the context
filter needs to be preserved.

This required some refactoring of the code responsible for making the
synchronization and was made more difficult by the fact that
experimental feature flags have to be loaded asynchronously and can
change over time.

Additionally there was an issue with React suspense when navigating from
a non-repo/results page to the repo/results page after a full page
reload, which is why I had to add `<Suspense>`.