insights: Fix admin ui queue items to use a compound key  (#47764)
It is possible for a single series to appear within multiple insights.
This change makes the ID of the queue item a compound key composed on
the backfill id and the unique id of the insight. The backfill ID is
required because that is the ID that all mutations are applied to. It is
not possible to retry, or reprioritize an insight or series, only the
backfill that has/will occur.

Also includes a sql fix for the queue position to properly sort the
backfills by cost, backfill_id.