experimental search input: Improve select and lang filters (#47756)
This PR includes a couple of improvements:

- Removed the padding from the CodeMirror line. Previously clicking on
the padding would position the cursor at the beginning of the line,
which was very confusing and seemed arbitrary to the user.
- Query caches are now initialized ones so that they persist across page
transitions.
- The empty state of the `lang:` filter now shows a small list of the
most common languages (better quality, less noise).
- The `select:` filter gets special treatment for value completion: By
default only the "top-level" values are shown (`repo`, `symbol`,
`commit`, etc) but as soon as the input matches one of the values and
the value has "sub-field" (e.g. `symbol`) all additional subfields are
rendered too. This keeps the initial list small and easy to process but
also surfaces the additional values to the user. In additional to that
we are smarter about when or when not to append a space after the value
to make discovering the additional values easier (see video). For
example, selecting `repo` will insert `repo ` (with space) allowing the
user to add a new filter/token, but selecting `symbol` will insert
`symbol` which in turn will surface all possible `symbol` subvalues.
- Reduced the number of results to show for repos and files. Still need
to tune that number so I guess it will continue to change for a while.