rbac: always upsert role assignment and permission assignment (#47780)
## Context
We have several `Assign`-ment methods in the store for `user_roles` and
`role_permissions`, however most of the call site for these methods
don't use the `types.UserRole` or `types.RolePermission` returned.
While working on #47700, I noticed that some tests were failing because:

* The tests had a call to `db.Users().Create`, which creates a site
admin user because the `GlobalState` hasn't been initialized. Then
calling `db.Users().SetIsSiteAdmin` after that results in a `sql: no
rows in result set` error being returned because the [`ON CONFLICT DO
NOTHING`
clause](https://sourcegraph.com/github.com/sourcegraph/sourcegraph/-/blob/internal/database/user_roles.go?L111)
in the upsert statement doesn't return a result set if the user is
already assigned a role.

This PR adds an extra check so that we don't return an error when an
upsert is done. I also updated the signature to only return an error
since we never use the relationship result in the call site for Role /
Permission assignment.