insights: admin UI graphql paging fix  (#47808)
The underlying paging controls made the assumption that any sort field
is both non-null and unique. This caused problems because we were
allowing sorting by fields that can be null (queue position) and state
which is not unique.

This change simplifies the paging cursor so that it will only contain
the backfill id then builds the where clause using that instead of the
relying on the underlying paging controller.

There does still remain the possibility that if a series were shared
between multiple insights and thus shared the same backfill and that
backfill fell at the border of a page, that paging could skip the
additional instances of that backfill. I consider this low ok for now
because at least one instance of the backfill will always appear and any
action on that will apply to all items.

## Test plan
Verify paging works for each sort option
verify search worked case insensitve