experimental search input: Add predicate support (#47809)
This PR updates the suggestion logic to render predicate suggestions
below value suggestions for the repo and file filters. For this I

- extended the suggestion extension to support "snippets", reusing
CodeMirror's built-in snippet support
- made the number of suggested repo/file values depending on whether
predicate results are present or not. This is to reduce the risk of
having the predicate section be scrolled out of view and to keep keep
the number of options manageable (we might want to limit the number of
predicates in an empty state too, I don't know yet)
- introduces constants for how many options should be rendered to make
them clearer and easier to change (we'll probably make this dynamic at
some point).