Writing perms to new table, behind a feature flag (#47159)
## Description

If feature flag `unified-user-repo-perms` is ON, we write to both the
new `user_repo_permissions` table and the old permission tables.

If feature flag is OFF, we keep the old behavior, only writing to the
old tables.

## Test plan

Manually tested locally
Integration test added
New behavior behind a feature flag, so no risk on affecting existing
instances.