app: add syntax highlighting support via Chroma (#47615)
Today, Sourcegraph App has no syntax highlighting. This is for the
reasons outlined in a new comment:

https://github.com/sourcegraph/sourcegraph/blob/96bcec929f3569c9e87755868e5458677eee8deb/cmd/frontend/internal/highlight/highlight.go#L452-L460

In this PR I pull in [Chroma](https://github.com/alecthomas/chroma), the
most popular Go syntax highlighter used by e.g. Hugo. It is effectively
a Go port of Pygments, and seems to have good compatibility with
Pygments grammars and decent performance. It supports [a decent chunk of
languages](https://github.com/alecthomas/chroma#supported-languages).

One challenge is that Chroma TokenTypes and SCIP SyntaxKinds do not
exactly agree / match 1:1. In some cases, Chroma is more detailed, while
in others SCIP is more detailed. In some cases Chroma has strange
choices, like e.g. when highlighting Go code "foo" in "package foo" is
considered a builtin. I did a best-effort mapping, and for most
languages it seems to work pretty well.

I don't want to pretend this is a perfect solution (is there ever such a
thing with syntax highlighting?); it's certainly not, and syntect is
definitely better - but I think this is *good enough* for now for App,
and it brings us from zero -> colors which is a clear win.

## Test plan

This change only affects App.

To test I manually reviewed a few handfuls of files:

Decent:

* Makefile
https://user-images.githubusercontent.com/3173176/218580862-25d9a9a2-fe02-404e-b90d-0172409286f1.png
* Go
https://user-images.githubusercontent.com/3173176/218580943-8b646a2f-8f65-49fd-bbd1-f9052e5239a9.png
* Typescript
https://user-images.githubusercontent.com/3173176/218582523-fbd60a5f-d43e-4aa3-bd85-2a4efce0b7bd.png
* Bazel
https://user-images.githubusercontent.com/3173176/218582689-d64cd875-3a46-4aa1-bf02-07c949cdef9b.png
* Nix
https://user-images.githubusercontent.com/3173176/218582776-a1cbb10e-f71f-4044-9999-b0d144f2f25d.png
* Python
https://user-images.githubusercontent.com/3173176/218582931-131f8737-87f0-463c-a121-117cd23667db.png
* C#
https://user-images.githubusercontent.com/3173176/218583003-6a8a5d55-3115-4c6d-a269-099407322378.png
* SCSS
https://user-images.githubusercontent.com/3173176/218583049-13010314-a139-4a7f-af6b-a1c8d26166c7.png
* Scala
https://user-images.githubusercontent.com/3173176/218583104-9d1877db-c9be-49b5-bb9f-6b8cea37b9f1.png
* C
https://user-images.githubusercontent.com/3173176/218583146-0fbf73f9-fd0a-4adf-8398-c8c4c2a902d9.png
* SQL
https://user-images.githubusercontent.com/3173176/218583191-555f9c9a-ae58-4ca8-a752-4543ab9bc460.png
* CSS
https://user-images.githubusercontent.com/3173176/218583242-c6c18357-01b4-4975-b23f-d19c04d565d7.png
* YAML
https://user-images.githubusercontent.com/3173176/218581046-e106444e-4626-4789-a967-de144f4c0199.png
* Dockerfile
https://user-images.githubusercontent.com/3173176/218581198-746d575f-ff84-45af-93c7-ee11150ea38d.png
* Plain text
https://user-images.githubusercontent.com/3173176/218581291-f40fef95-8694-48ae-b41f-78c05287f32d.png
* ps1
https://user-images.githubusercontent.com/3173176/218581520-2ade3d56-5915-4b31-a202-86831941fcfb.png
* Makefile
https://user-images.githubusercontent.com/3173176/218581779-e64bce10-8c37-425d-a5e4-fdded98b9f1a.png
* Shell
https://user-images.githubusercontent.com/3173176/218581911-8a0d9753-4f97-47c5-a2e0-dde2576e9968.png

Poor:

* JSON
https://user-images.githubusercontent.com/3173176/218581465-2894fa9c-7ca1-4336-855c-cf602a22d343.png
* Some YAML files (but not all)?
* go.mod
https://user-images.githubusercontent.com/3173176/218582069-838f7ce1-6736-4a4b-aec8-d21dbe339596.png

No highlighting:

* Markdown
* Procfile
* .patch

---------

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>