return nil payload when LocalCodeIntel fails
Our client code expects a nil payload when LocalCodeIntel fails. This
seems to be used as a sentinel to fall back to search-based code intel
in the client code. For the gRPC implementation, we propagated any
errors we received, but this actually ends up breaking things because
errors cause the client to not fall back to search-based code intel and
instead throw the error.