client/web: Add UI elements for Azure DevOps auth provider (#47865)
This commit does a few related things:

1. Add an icon on the signin page for Azure DevOps

2. Implement external data functions and add code to show the connected
account data for Azure DevOps

3. Finally, it fixes a bug (discovered as a result of implementing this)
with the `parseProvider` where I was incorrectly using the `Kind`
instead of `Type`.

Sidenote: Since both are strings and the only difference is that the
former is in `UPPERCASE` while the latter is in `lowercase`, this is an
easy mistake to make. The real fix should be that these should be custom
types so that methods that accept these as an argument get the powers of
type checking. Expect a follow up PR.