bzl: avoid running test with external network reqs
External HTTP requests are not compatible with running tests fully
sandboxed as bazel expects them. This PR add a quick check to avoid
running them when Bazel is the test runner.

On the long term, I might consider adding a bunch a helpers to deal with
this as we progress through various edge cases.

For now that will do.

Test plan: tests passed with Bazel, locally.