Gitserver: simplify signature of AddrForRepo (#47920)
Currently, AddrForRepo takes a context and returns an error. The context
is never used and the error is always nil. This updates the signature of
the method to reflect that.

Setup for some work to generalize address picking for gRPC stuff.

Maybe someday `AddrForRepo` won't be exposed at all...

## Test plan

Existing tests. Semantics-preserving, just removing unused variables.