Storm 🌪️: Preload logo via <link rel="preload"> (#47838)
Digging through the resource water-falling of our navbar component, I
noticed that we start to load the logo relatively late, even though it's
shown on any route. Additionally, we use the same SVG as the favicon but
with a slightly different URL causing the browser to have to make
another request:

<img width="1942" alt="Screenshot 2023-02-20 at 15 45 11"
src="https://user-images.githubusercontent.com/458591/220137301-37ba262d-498d-4bf3-9554-3e7ea218b4f3.png">

This PR adds a way for the server-generated HTML to include `<link
rel="preload">` elements to preload the .svg in question.

It also fixes the favicon to now use the same URL. It does seem that
`<link rel="preload">` is not used by the browser for the favicon but
the request might still be cached with the right cache headers (and if
that also doesn't apply, at least a CDN might be able to do that!)

The lazy loading nature of the image also caused it to flicker during
page loads (this video is from dotcom):

https://user-images.githubusercontent.com/458591/220139052-355448d7-3a7a-4d8e-bb78-ea61dca7b8dc.mov

With preloading, it appears instantly. 💥