feature flags: Clean up (and fix) deprectated search-related feature flags
These flags have been removed in #46086 and #46045 but were
accidentally(?) added back to the schema in (#46086).

Furthermore the `showSearchContext` flag was still accessed in a couple
of places, resulting in a broken experience if it was set to `false`
(see screenshot).

This PR removes the flags and removes all uses of `showSearchContext`
flag.