Fix code intel hover navigation propagation issue
PR #47595 introduced an issue where link navigations inside the code
mirror components are not propagated to the main React router context.

This is caused by the fact that these components are rendered into their
own React roots which are wrapped with a custom router context that no
longer communicates with the main context.

To fix this, I've added a small wrapper that listens for route changes
in the parent and propagates it to the main router via the navigate
function.