Gitserver: centralize address resolution in `GitserverAddresses` (#47924)
This is some more setup for gRPC so we can have avoid opening a
connection per RPC call.

Previously, a gitserver client had a dependency on two different
closures: one to fetch the list of gitserver addresses, and one to fetch
the set of repos that are pinned to a specific gitserver instance. We
had quite a few methods and functions that expose this information (with
some duplication on the server side), and no place that we could
atomically swap out a set of associated gRPC connections.

This centralizes the address resolution interface around
`GitServerAddresses`, which already existed, but was not broadly used.
This will allow us to open a global set of gRPC connections so we no
longer have open connections per-RPC.

## Test plan

Updated tests. No behavior change, so I'm relying on existing tests
here.
-->