experimental search input: Fix overflow/line wrapping behavior
Currently the input wraps around for very long queries, but it doesn't
look great because:
- The mode button + separator only cover the first line
- Lines have no padding and look "squished" togehter.

This PR changes the implementation to replicate the behavior of the
current search input: scroll overflow with hidden scrollbar.

To make the scroll behavior work the mode button/indicator had to be
reimplemented, otherwise it would have been scrolled out of view. The
button is now implemented in React directly which has a couple of
advantages:

- It's now in "tab order"
- We can use the Tooltip component from Wildcard

This PR also fixes some dark mode related issues (but not all of them).
Specifically I misunderstood the purpose of the `{dark: true}` option
passed to `EditorView.theme` (which causes the CodeMirror to think dark
mode is active which also affects the text selection style). This change
fixes that.